Give host model fields all values after added to db
