+++ MultiplayerWebFramework/HostServicesAPI/Interfaces/ICluster.cs
@@ -12,7 +12,7 @@ namespace HostServicesAPI.Interfaces
     public interface ICluster
     {
         public List<GameInstanceModel> ActiveGameInstances { get; set; }
-        public Task<HttpResponseMessage> SpinUpGameInstance(Game game, string port, string args, string filePath);
+        public Task<HttpResponseMessage> SpinUpGameInstance(Game game, string port, string args, int hostId, string filePath);
         public Task<HttpResponseMessage> ShutDownGameInstance(int inGameInstanceId);
     }
 }
