+++ MultiplayerWebFramework/HostServicesAPI/Objects/SetupTeardownHostedService.cs
@@ -5,8 +5,10 @@ using MWFModelsLibrary.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Net.Http;
 using System.Net.Http.Json;
+using System.Net.Sockets;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
@@ -47,7 +49,7 @@ namespace HostServicesAPI.Objects
         // We can return void asynchronously since this is a callback
         private async void OnStarted()
         {
-            applicationHostModel = new HostModel { HostIp = "d", HostServicesAPISocketAddress = "34243:5245", IsActive = true };
+            applicationHostModel = new HostModel { HostIp = GetMachineIP(), HostServicesAPISocketAddress = GetMachineIP(), IsActive = true };
 
             HttpResponseMessage httpResponse = null;
             try
@@ -55,7 +57,6 @@ namespace HostServicesAPI.Objects
                 _logger.Log(LogLevel.Information, "Attempting to register us with the database (creating db host entry)");
                 HttpClient client = _clientFactory.CreateClient("MWFHostServicesAPIClient");
                 httpResponse = await client.PostAsJsonAsync<HostModel>(@"http://localhost:7071/api/CreateHostAndReturnId", applicationHostModel, new JsonSerializerOptions() { PropertyNameCaseInsensitive = true });
-
             }
             catch (Exception e)
             {
@@ -72,6 +73,8 @@ namespace HostServicesAPI.Objects
                 else if (httpResponse?.IsSuccessStatusCode == true)
                 {
                     _logger.Log(LogLevel.Information, "Successful status code: " + httpResponse.StatusCode.ToString() + "\nHost added to database! API is now ready for requests!");
+                    int id = await HttpContentJsonExtensions.ReadFromJsonAsync<int>(httpResponse.Content);
+                    applicationHostModel.Id = id;
                 }
             }
 
@@ -89,6 +92,22 @@ namespace HostServicesAPI.Objects
             // After all game instaces from this host are shutdown and removed from the database, we must remove the host model from the database (must happen after since removing the host model before could result in a rejection since there may be forign keys from Game Instances referenceing it)
             // Also how do we want to handle cases where removing game instance from database fails? Should we not then remove the host from the database at all since this leaves a possibility the host won't be removed? Should we maybe add to the stored procedure for removing a host to remove all game instances that have forien keys to it? 
             /*var result = Http.PostAsJson<GameInstanceModel>(@"http://localhost:7071/api/RemoveHostById", newGameInstanceToAdd, new JsonSerializerOptions() { PropertyNameCaseInsensitive = true });*/
+        }
+
+
+
+
+        private static string GetMachineIP()
+        {
+            string localIp;
+            using (Socket socket = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, 0))
+            {
+                socket.Connect("8.8.8.8", 65530);
+                IPEndPoint endPoint = socket.LocalEndPoint as IPEndPoint;
+                localIp = endPoint.Address.ToString();
+            }
+
+            return localIp;
         }
     }
 }
